+++ app.py
@@ -1,16 +1,5 @@
 from flask import Flask, render_template, request, redirect
-
-
-class Foods:
-    def __init__(self):
-        self.food = ''
-
-    def add_food(self, food):
-        self.food = food
-
-    def get_food(self):
-        return self.food
-
+from model import Foods
 
 app = Flask(__name__)
