+++ model.py
@@ -1,0 +1,9 @@
+class Foods:
+    def __init__(self):
+        self.food = ''
+
+    def add_food(self, food):
+        self.food = food
+
+    def get_food(self):
+        return self.food
