+++ static/styles/index.css
@@ -1,0 +1,7 @@
+body {
+ margin: 10px 10px 0px 0px;
+ padding: 20px 0px 0px 20px;
+}
+.container {
+  text-align: center;
+}
