+++ templates/index.html
@@ -1,17 +1,22 @@
 <!DOCTYPE html>
 <html>
 <head>
+    <meta charset="utf-8">
     <title>Group Three Foods</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <link rel="stylesheet" href="{{ url_for('static',filename='styles/index.css') }}" />
 </head>
 <body>
-    <h1>Group Three Foods</h1>
-    <hr>
-    <form action="/item" method="POST">
-        <label for="food">Enter a Dish:</label><br>
-        <input type="text" id="food" name="food"><br>
-        <input type="submit" value="Submit">
-    </form>
-    <br>
-    <p>{{food_response}}</p>
+    <div class="container">
+        <h1>Group Three Foods</h1>
+        <hr>
+        <form action="/item" method="POST">
+            <label for="food">Enter a Dish:</label>
+            <input type="text" id="food" name="food">
+            <input type="submit" value="Submit">
+        </form>
+        <br>
+        <p>{{food_response}}</p>
+    </div>
 </body>
 </html>
